nvme/pci: Remove last_sq_tail
We don't allocate enough tags for the submission queue tail to pass the
head. Remove tracking for this condition, and update code comments
accordingly. The driver's .commit_rqs() function can unconditionally
ring the doorbell since it is only called if the sq_tail had previously
moved, but was not rung.

Reviewed-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Keith Busch <kbusch@kernel.org>